Add composite primary key to followers table (#46)
* Add composite key to followers table

Because the Entity Framework needs the tables to have primary key in
order to be able to track changes then we have decided to add this to
our followers table. Also because logically a user can only follow
another user once.

Co-authored-by: Amalie Bøgild Sørensen <malleboegild@gmail.com>

* Make primary key in follower entity model to not allow null

Co-authored-by: Amalie Bøgild Sørensen <malleboegild@gmail.com>

* Change raw sql queries to native EF C# sql queries

Due to missing composite primary key in the followers table EF was not
able to track changes. Now it is added so we should be able to handle
that now.

Co-authored-by: Amalie Bøgild Sørensen <malleboegild@gmail.com>

---------

Co-authored-by: Amalie Bøgild Sørensen <malleboegild@gmail.com>